Move GFNI tests to amd64
